Update patch set 1
Patch Set 1: Code-Review-2

(6 comments)

The path change won't really make a difference.

Could you retest on a fresh live usb from a fresh folder, and verify that /etc/udev/rules.d is empty before trying to use it.

If that works then we know it was the order (which seems strange but I'll take it)

If not, we'll see.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@35266860-6252-469c-a67d-185064479fc0>
Label: Code-Review=-2