Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Just putting this out here for discussion.

I still vote for no vendor directory. It adds so much bloat to the repo and go mod does a good job with reproducible builds assuming dependencies don't go away. There is some level of protection in the fact that go proxy defaults to https://proxy.golang.org/ now for dependencies, but it still doesn't completely guarantee a package won't go away. If we are particularly worried about dependencies going away, it's possible to setup our own go proxy that could permanently store dependencies without needing it to be in the git repo.

Patch-set: 1
CC: Gerrit User 1000006 <1000006@35266860-6252-469c-a67d-185064479fc0>