Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Just putting this out here for discussion.
> 
> I still vote for no vendor directory. It adds so much bloat to the repo and go mod does a good job with reproducible builds assuming dependencies don't go away. There is some level of protection in the fact that go proxy defaults to https://proxy.golang.org/ now for dependencies, but it still doesn't completely guarantee a package won't go away. If we are particularly worried about dependencies going away, it's possible to setup our own go proxy that could permanently store dependencies without needing it to be in the git repo.

Agree that is does bloat up the repo, but size wise it's not too big since it's all source code.

I prefer this since it's simple, and should always work.
A package can go away, or it can be replaced by something malicious - that has happened.

I don't mind having this in a separate repo somehow, but looks like it'd still be additional complication, whereas this is supported out of the box with just one command, making updates easy too (just run `go mod vendor` and it'd be done and known to work.)

Patch-set: 1